Extract functions from spec into desalinate
Working title, but this coud become the test utility library for the
framework, if it turns out to be useful. Avoid annotating some concise
arrow functions to avoid tcomb typechecking unnecessarily. Also adds
Disposer to a global `declare`, since it's used with Kefir's `stream`.
Also tells Flow that mocha supports returning Observables, even though
that hasn't landed yet.
