+++ .babelrc
@@ -1,6 +1,12 @@
 {
   "plugins": [
-    "tcomb",
+    ["tcomb", {
+      "globals": [
+        {
+          "Disposer": true
+        }
+      ]
+    }],
     "ramda",
     "transform-object-rest-spread",
     "syntax-flow",
