+++ client/type/index.js
@@ -2,5 +2,4 @@
 export * from './action';
 export * from './ajax';
 export * from './state';
-export * from './kefir';
 export * from './tinymce';
