--- client/type/kefir.js
@@ -1,2 +1,0 @@
-// @flow
-export type Disposer = () => void;
