+++ flow-typed/npm/mocha_v3.1.x.js
@@ -1,7 +1,8 @@
 // flow-typed signature: 6b82cf8c1da27b4f0fa7a58e5ed5babf
 // flow-typed version: edf70dde46/mocha_v3.1.x/flow_>=v0.22.x
+import type { Observable } from 'kefir';
 
-type TestFunction = ((done: () => void) => void | Promise<mixed>);
+type TestFunction = ((done: () => void) => void | Promise<mixed> | Observable<mixed>);
 
 declare var describe : {
     (name:string, spec:() => void): void;
